% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet.R
\name{as_facet_layer}
\alias{as_facet_layer}
\title{Convert Object to ggplot2 Facet Layer}
\usage{
as_facet_layer(x, ...)
}
\arguments{
\item{x}{Object to convert, typically a \code{Facet} object created by \code{facet()}}

\item{...}{Additional arguments passed to underlying ggplot2 facet functions}
}
\value{
A ggplot2 facet layer (either \code{facet_wrap()} or \code{facet_grid()}) that can be
added to a ggplot object
}
\description{
Generic function for converting facet specification objects into ggplot2 facet layers.
This allows seamless integration of custom facet objects with ggplot2 visualizations.
}
