% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{graphics}
\alias{graphics}
\title{Create a List of Quosures for Graphics Layers}
\usage{
graphics(...)
}
\arguments{
\item{...}{Additional ggplot2 layer expressions to be captured as quosures}
}
\value{
A list of quosures containing the unevaluated expressions
}
\description{
Helper function to capture additional ggplot2 layers as quosures. This is useful
for passing extra graphics elements to plotting functions.
}
