% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{is_data_output}
\alias{is_data_output}
\title{Check if object is a DataOutput}
\usage{
is_data_output(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
A logical value indicating whether the object inherits from DataOutput
}
\description{
Check if object is a DataOutput
}
