% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet.R
\name{is_facet_class}
\alias{is_facet_class}
\title{Check if Object is a Facet Specification}
\usage{
is_facet_class(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
\code{TRUE} if \code{x} is a \code{Facet} object, \code{FALSE} otherwise
}
\description{
Check if Object is a Facet Specification
}
