% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{is_plot_output}
\alias{is_plot_output}
\title{Check if object is a PlotOutput}
\usage{
is_plot_output(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
A logical value indicating whether the object inherits from PlotOutput
}
\description{
Check if object is a PlotOutput
}
