% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{is_styles}
\alias{is_styles}
\title{Check if an object is a styles collection}
\usage{
is_styles(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
A logical scalar indicating whether \code{x} is a styles object
}
\description{
Check if an object is a styles collection
}
