% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{new_styles}
\alias{new_styles}
\title{Create a collection of styles}
\usage{
new_styles(...)
}
\arguments{
\item{...}{Style formulas, each specified as \code{factors ~ aesthetics}}
}
\value{
A styles object containing the specified style definitions
}
\description{
Create a collection of styles
}
