% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nouel2026MouseBehavior-package.R
\docType{package}
\name{Nouel2026MouseBehavior-package}
\alias{Nouel2026MouseBehavior}
\alias{Nouel2026MouseBehavior-package}
\title{Nouel2026MouseBehavior: A package for analyzing mouse behavior data from Liz Nouel's paper}
\description{
This package provides tools for analyzing and visualizing mouse behavior data as presented in Liz Nouel's 2026 paper on mouse behavioral patterns.
}
\author{
\strong{Maintainer}: Kyle Schaubroeck \email{kyle.schaubroeck@wisc.edu} (\href{https://orcid.org/0009-0001-7011-4041}{ORCID})

}
\keyword{internal}
