% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{responses}
\alias{responses}
\title{Create Response Variable Expressions}
\usage{
responses(...)
}
\arguments{
\item{...}{Named expressions representing response variables. At least one
response must be provided.}
}
\value{
A named list of unevaluated expressions.
}
\description{
Creates a named list of response variable expressions for use in model formulas.
All arguments must be named and will be captured as expressions.
}
