% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{scale_x_discrete_style}
\alias{scale_x_discrete_style}
\title{Apply discrete x-axis scale from a style theme}
\usage{
scale_x_discrete_style(theme, keys, ...)
}
\arguments{
\item{theme}{A styles object}

\item{keys}{Factor names to match and sort by}

\item{...}{Additional arguments passed to \code{ggplot2::scale_x_discrete()}}
}
\value{
A ggplot2 scale object
}
\description{
Apply discrete x-axis scale from a style theme
}
