% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{aesthetic}
\alias{aesthetic}
\title{Aesthetic}
\usage{
aesthetic(
  ...,
  shape = NA_integer_,
  color = NULL,
  colour = NULL,
  fill = color,
  linewidth = NA_real_,
  linetype = NA_character_,
  label = NA_character_
)
}
\value{
An object of class \code{Aesthetic}.
}
\description{
S7 class describing graphical aesthetics used in styles.
}
