% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet.R
\name{as_facet_layer}
\alias{as_facet_layer}
\title{Generic: Convert to facet layer}
\usage{
as_facet_layer(x, ...)
}
\arguments{
\item{x}{Object to convert.}
}
\value{
A ggplot2 facet layer or an object understood by ggplot2.
}
\description{
Create a generic function for converting objects to a ggplot2 facet layer.
}
