% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{scale_fill_style}
\alias{scale_fill_style}
\title{Apply fill scale from a style theme}
\usage{
scale_fill_style(theme, keys, ...)
}
\arguments{
\item{theme}{A styles object}

\item{keys}{Factor names to match and sort by}

\item{...}{Additional arguments passed to \code{ggplot2::scale_fill_manual()}}
}
\value{
A ggplot2 scale object
}
\description{
Apply fill scale from a style theme
}
